animation/extract hooks from TierAnimationSection
